% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{anova,lgc-method}
\alias{anova,lgc-method}
\title{Comparing the fit of LGC objects.}
\usage{
\S4method{anova}{lgc}(object, ...)
}
\arguments{
\item{object}{\code{lgc} object. An \code{lgc} object to be compared against other \code{lgc} objects.}

\item{...}{\code{lgc} object. More \code{lgc} objects to be compared.}
}
\description{
Comparing the fit of LGC objects.
}
