% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{anova.Lgc}
\alias{anova.Lgc}
\title{Compares the lavaan objects contained in Lgc objects.}
\usage{
\method{anova}{Lgc}(object, ...)
}
\arguments{
\item{object}{Object of class Lgc.}

\item{...}{Additional arguments passed down to
anova(lavaan_object, ...).}
}
\description{
Compares the lavaan objects contained in Lgc objects.
}
