% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrastPlot.R
\name{ContrastPlot}
\alias{ContrastPlot}
\title{ContrastPlot class}
\description{
Some description
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContrastPlot-new}{\code{ContrastPlot$new()}}
\item \href{#method-ContrastPlot-plot}{\code{ContrastPlot$plot()}}
\item \href{#method-ContrastPlot-clone}{\code{ContrastPlot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastPlot-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastPlot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastPlot$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastPlot-plot"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastPlot-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastPlot$plot(
  lgc,
  hypothesis,
  geom = c("line", "point"),
  interval = c("none", "sd", "ci")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContrastPlot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContrastPlot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContrastPlot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
