% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dag.R
\name{dag}
\alias{dag}
\title{Wrapper method for Dag$new()}
\usage{
dag(lgc, ...)
}
\arguments{
\item{lgc}{Object of class Lgc.}

\item{...}{Arguments passed down to dag$plot().}
}
\description{
Wrapper method for Dag$new()
}
