% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Simple data simulation function for test cases.}
\usage{
get_test_data(n = 100L)
}
\arguments{
\item{n}{Integer. Number of cases simulated.}
}
\description{
Simple data simulation function for test cases.
}
