% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypothesis.R
\name{Hypothesis}
\alias{Hypothesis}
\title{Hypothesis class}
\description{
Some description
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_L}}{Return the L matrix of the hypothesis.}

\item{\code{get_M}}{Return the M matrix of the hypothesis.}

\item{\code{get_P}}{Return the P matrix of the hypothesis.}

\item{\code{get_description}}{Return the description of the hypothesis.}

\item{\code{expose}}{Returns the private structure of the object. For debugging purposes only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Hypothesis-new}{\code{Hypothesis$new()}}
\item \href{#method-Hypothesis-get_wald_string}{\code{Hypothesis$get_wald_string()}}
\item \href{#method-Hypothesis-get_contrasts}{\code{Hypothesis$get_contrasts()}}
\item \href{#method-Hypothesis-get_contrasts_old}{\code{Hypothesis$get_contrasts_old()}}
\item \href{#method-Hypothesis-clone}{\code{Hypothesis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hypothesis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Hypothesis-new}{}}}
\subsection{Method \code{new()}}{
Validates the input for different functions.


Hypothesis class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hypothesis$new(M = NULL, L = NULL, P = NULL, description = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{M}}{M matrix. Can be omitted if the number of dependent
variables equals 1.}

\item{\code{L}}{L matrix. Can be omitted if the number of groups equals 1.}

\item{\code{P}}{P matrix. Can be omitted if the number of covariates
equals 0.}

\item{\code{description}}{Character. Description of the hypothesis. Can be
any character.}

\item{\code{lgc}}{Object of class Lgc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hypothesis-get_wald_string"></a>}}
\if{latex}{\out{\hypertarget{method-Hypothesis-get_wald_string}{}}}
\subsection{Method \code{get_wald_string()}}{
Returns the a string for the Wald test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hypothesis$get_wald_string(lgc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lgc}}{Object of class Lgc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hypothesis-get_contrasts"></a>}}
\if{latex}{\out{\hypertarget{method-Hypothesis-get_contrasts}{}}}
\subsection{Method \code{get_contrasts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hypothesis$get_contrasts(lgc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hypothesis-get_contrasts_old"></a>}}
\if{latex}{\out{\hypertarget{method-Hypothesis-get_contrasts_old}{}}}
\subsection{Method \code{get_contrasts_old()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hypothesis$get_contrasts_old(lgc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Hypothesis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Hypothesis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hypothesis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
