% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{Lgc}
\alias{Lgc}
\alias{lgc}
\alias{lgc_dummy}
\title{Prints the reliabilities of the indicators of latent variables.}
\usage{
lgc(...)

lgc_dummy(...)
}
\description{
Some description
}
\details{
Lgc Class

\code{lgc()} is a wrapper for the R6 class constructor
\code{Lgc$new()}. See \code{Lgc$new()} below for a list of arguments.
}
\keyword{internal}
\keyword{internal#'}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_par_table}}{Read only. Retrieves the parameter table as tibble.}

\item{\code{get_group_labels}}{Read only. Retrieves the group_labels.}

\item{\code{expose}}{Read only. Returns the private structure of the object. For debugging purposes only.}

\item{\code{get_hypotheses}}{Read only. Returns the list of hypotheses. For debugging purposes only.}

\item{\code{get_par_labels}}{Read only. Returns an array of parameter labels used in the lavaan syntax.}

\item{\code{get_sem_object}}{Read only. Returns the estimated lavaan SEM object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Lgc$new()}}
\item \href{#method-specify}{\code{Lgc$specify()}}
\item \href{#method-specify_dummy}{\code{Lgc$specify_dummy()}}
\item \href{#method-estimate}{\code{Lgc$estimate()}}
\item \href{#method-get_lav_syntax}{\code{Lgc$get_lav_syntax()}}
\item \href{#method-get_estimates}{\code{Lgc$get_estimates()}}
\item \href{#method-test}{\code{Lgc$test()}}
\item \href{#method-tests_table}{\code{Lgc$tests_table()}}
\item \href{#method-print}{\code{Lgc$print()}}
\item \href{#method-summary}{\code{Lgc$summary()}}
\item \href{#method-compare}{\code{Lgc$compare()}}
\item \href{#method-plot}{\code{Lgc$plot()}}
\item \href{#method-get_data}{\code{Lgc$get_data()}}
\item \href{#method-set_data}{\code{Lgc$set_data()}}
\item \href{#method-power_analysis}{\code{Lgc$power_analysis()}}
\item \href{#method-reliabilities}{\code{Lgc$reliabilities()}}
\item \href{#method-clone}{\code{Lgc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Returns a tibble that can be used to plot a DAG.


Lgc class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{labels}}{Character. Label type on the edged.}

\item{\code{digits}}{Integer. How many digits should be printed?}

\item{\code{...}}{Parameters passed to get_estimates().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-specify"></a>}}
\if{latex}{\out{\hypertarget{method-specify}{}}}
\subsection{Method \code{specify()}}{
Specifies an Lgc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$specify(
  data = NULL,
  etas,
  c_matrix_within,
  resid_cov = list(),
  equal_resid_cov = list(),
  invariance_within = "strong",
  compound_symmetry = FALSE,
  sphericity = list(),
  c_matrix_between = TRUE,
  invariance_between = "strong",
  covariates = NULL,
  fixed_covariates = TRUE,
  lv_scaling = "effect",
  group = NULL,
  group_labels = NULL,
  group_weights = "fixed",
  variance_homogeneity = FALSE,
  hypotheses = NULL,
  append = NULL,
  dummy = FALSE,
  verbose = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{tibble. Will be converted to tibble if not already a tibble.}

\item{\code{etas}}{Named list of character vectors of length q. Each list
element (character vector) represents an eta variable. The elements
of the character vector are manifest variables (i.e., indicators)
from the data set that measure the corresponding eta variable. The
list names correspond to the names of the eta variables.
E.g.:
\code{etas = list(
    eta1 = c("Y11", "Y12"),
    eta2 = c("Y21", "Y22"),
    eta3 = c("Y31", "Y32")
)}}

\item{\code{c_matrix_within}}{Numeric q x q matrix. Contrast matrix for the
within-subjects design. Must be a square matrix.
E.g.:
\code{
c_matrix_within = matrix(c(
    1/3, -sqrt(1/2),  sqrt(1/6),
    1/3,          0, -sqrt(2/3),
    1/3,  sqrt(1/2),  sqrt(1/6)
), nrow = 3)}}

\item{\code{resid_cov}}{List of character vectors. Each character vector
contains manifest variables from the data set among which residual
covariances should be implemented. Residual covariances can vary.}

\item{\code{equal_resid_cov}}{List of character vectors. Each character vector
contains manifest variables from the data set among which residual
covariances should be implemented. Residual covariances are
constrained to be equal.}

\item{\code{invariance_within}}{Logical. Default is TRUE. Indicates whether
strong measurement invariance across latent dependent variables eta
should be imposed.}

\item{\code{compound_symmetry}}{Logical. Indicates whether compound symmetry
should be imposed.}

\item{\code{sphericity}}{List of character vectors. Each character vector
contains indices of pi variables among which sphericity should be
imposed.}

\item{\code{c_matrix_between}}{Numeric p x p matrix. Contrast matrix for the
between-subjects design. Must be a square matrix.
E.g.:
\code{
c_matrix_between = matrix(c(
    1/3, -sqrt(1/2),  sqrt(1/6),
    1/3,          0, -sqrt(2/3),
    1/3,  sqrt(1/2),  sqrt(1/6)
), nrow = 3)}}

\item{\code{invariance_between}}{Logical. Default is \code{TRUE}.
Indicates whether measurement invariance across groups should be
imposed.}

\item{\code{covariates}}{Named list of character vectors. Same format as the
argument etas.}

\item{\code{fixed_covariates}}{Logical. Indicates whether manifest
covariates should be treated as fixed.}

\item{\code{lv_scaling}}{Character. Possible values:
\code{c("effect", "referent")}. Default is "effect". Scheme to
identify the latent variable scale. For \code{"effect"}: effect coding
scheme freely estimates all loadings and intercepts of the indicators
and constraints the average of the loadings to 1 and the average of
the intercepts to 0. For \code{"referent"}: referent indicator scheme
sets the first loading to 1 and the first intercept to 0.}

\item{\code{group}}{Character. Contains the grouping variable from the data
set.}

\item{\code{group_labels}}{Character vector of length p. Contains unique group
names of the group variable from the data set. Optional, default is
NULL. If not supplied, labels will be pulled from the data set using
the provided group variable.
E.g.: \code{groups = c("group1", "group2", "group3")}}

\item{\code{group_weights}}{Either numeric vector of length p (same length
as \code{group_labels}) or one of the characters
\code{c("fixed", "stochastic")}. Default is \code{"fixed"}.}

\item{\code{variance_homogeneity}}{Logical. Default is \code{TRUE}.
Indicates whether homogeneity of variance should be imposed across
groups.}

\item{\code{hypotheses}}{List of objects of class Hypothesis.}

\item{\code{append}}{Character. lavaan syntax that is to be appended to the
model string.}

\item{\code{dummy}}{Logical. Indicates whether the model should be
estimated.}

\item{\code{verbose}}{Logical.}

\item{\code{...}}{Additional arguments passed down to lavaan.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-specify_dummy"></a>}}
\if{latex}{\out{\hypertarget{method-specify_dummy}{}}}
\subsection{Method \code{specify_dummy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$specify_dummy(
  n_etas = 1,
  n_indicators_eta = 1,
  n_groups = 1,
  c_matrix_within = NULL,
  n_covariates = NULL,
  hypotheses = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-estimate}{}}}
\subsection{Method \code{estimate()}}{
Estimates the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$estimate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_lav_syntax"></a>}}
\if{latex}{\out{\hypertarget{method-get_lav_syntax}{}}}
\subsection{Method \code{get_lav_syntax()}}{
Retrieves the lavaan syntax.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$get_lav_syntax(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-get_estimates}{}}}
\subsection{Method \code{get_estimates()}}{
Returns the point estimates of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$get_estimates(
  what = c("variance", "covariance", "intercept", "measurement", "resid_var",
    "resid_cov", "regression"),
  only_unique_labels = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{Character vector.}

\item{\code{only_unique_labels}}{Logical. Default is FALSE.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test"></a>}}
\if{latex}{\out{\hypertarget{method-test}{}}}
\subsection{Method \code{test()}}{
Tests the provided hypothesis on the lgc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$test(hypothesis)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypothesis}}{Object of class Hypothesis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tests_table"></a>}}
\if{latex}{\out{\hypertarget{method-tests_table}{}}}
\subsection{Method \code{tests_table()}}{
Tests multiple hypotheses on the lgc object and returns
the results as a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$tests_table(hypotheses)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypothesis}}{List of objects of class Hypothesis.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints the lavaan object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Prints a summary for the Lgc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$summary(detailed = F, fit_measures = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{detailed}}{Logical. Indicates whether a detailed summary
should be printed.}

\item{\code{fit_measures}}{Logical. Indicates whether fit measures should be
printed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare"></a>}}
\if{latex}{\out{\hypertarget{method-compare}{}}}
\subsection{Method \code{compare()}}{
Compares the lavaan objects contained in Lgc objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$compare(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed down to
anova(lavaan_object, ...).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots a directed acyclic graph for the Lgc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$plot(
  label = "both",
  what = c("regression", "measurement"),
  variance.direction = 90,
  variance.span = 75,
  variance.strength = 1.5,
  resid_var_left.direction = 180,
  resid_var_left.span = -75,
  resid_var_left.strength = 1.5,
  resid_var_right.direction = 0,
  resid_var_right.span = 75,
  resid_var_right.strength = 1.5,
  covariance.strength = 0.1,
  transparent_nodes = FALSE,
  only_first_group = FALSE,
  pi.x = NULL,
  eta.x = NULL,
  eta_indicator.x = NULL,
  covariate.x = NULL,
  covariate_indicator.x = NULL,
  left_intercept.x = NULL,
  right_intercept.x = NULL,
  left_residual.x = NULL,
  right_residual.x = NULL,
  node_distance_scale.x = 1,
  node_distance_scale.y = 1,
  vjust = -0.5,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to get_tidy_dagitty().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Returns the data tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_data}{}}}
\subsection{Method \code{set_data()}}{
Sets the data tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{tibble.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-power_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-power_analysis}{}}}
\subsection{Method \code{power_analysis()}}{
Plots a directed acyclic graph for the Lgc object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$power_analysis(data_syntax, sample_size = 100L, replications = 500L, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_syntax}}{List of characters. Each character represents
syntax to simulate data for a separate group.}

\item{\code{sample_size}}{Vector of integers. Must be of length or the
same length as data_syntax.}

\item{\code{replications}}{Integer. Number of replications.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reliabilities"></a>}}
\if{latex}{\out{\hypertarget{method-reliabilities}{}}}
\subsection{Method \code{reliabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$reliabilities()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lgc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
