% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lgc.R
\name{plot.Lgc}
\alias{plot.Lgc}
\title{Wrapper method for lgc$plot()}
\usage{
\method{plot}{Lgc}(x, ...)
}
\arguments{
\item{x}{Object of class Lgc.}

\item{...}{Arguments passed down to lgc$plot().}
}
\description{
Wrapper method for lgc$plot()
}
