% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{plot.Power}
\alias{plot.Power}
\title{Wrapper method for power$plot()}
\usage{
\method{plot}{Power}(x, ...)
}
\arguments{
\item{x}{Object of class Power.}

\item{...}{Arguments passed down to power$plot().}
}
\description{
Wrapper method for power$plot()
}
