% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{power_analysis_lgc}
\alias{power_analysis_lgc}
\title{Wrapper method for Power$get_power_semnova_script()}
\usage{
power_analysis_lgc(...)
}
\arguments{
\item{...}{Additional arugments passed to Power$get_power_semnova_script().}
}
\description{
Wrapper method for Power$get_power_semnova_script()
}
