% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{power_script_lgc}
\alias{power_script_lgc}
\title{Wrapper method for Power$get_power_lgc_script()}
\usage{
power_script_lgc(...)
}
\arguments{
\item{...}{Additional arugments passed to Power$get_power_lgc_script().}
}
\description{
Wrapper method for Power$get_power_lgc_script()
}
