% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{print.Power}
\alias{print.Power}
\title{Wrapper method for power$print()}
\usage{
\method{print}{Lgc}(x, ...)
}
\arguments{
\item{x}{Object of class Power.}

\item{...}{Arguments passed down to power$print().}
}
\description{
Wrapper method for power$print()
}
