% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{semnova_test_data}
\alias{semnova_test_data}
\title{This data set serves for examples and tests.}
\format{
A data frame with 100 rows and 6 variables:
}
\usage{
semnova_test_data
}
\description{
This is a simulated data set that 100 observation of six normally
distributed variables with mean = 0, variance = 1 and covariance 0.5.
}
\keyword{datasets}
