% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Effect.R
\name{SemnovaEffect}
\alias{SemnovaEffect}
\alias{effSources.SemnovaEffect}
\alias{formula.SemnovaEffect}
\alias{family.SemnovaEffect}
\alias{model.frame.SemnovaEffect}
\alias{model.offset.SemnovaEffect}
\alias{model.matrix.SemnovaEffect}
\alias{expand.model.frame.SemnovaEffect}
\alias{getCall.SemnovaEffect}
\alias{vcov.SemnovaEffect}
\alias{coefficients.SemnovaEffect}
\alias{coef.SemnovaEffect}
\alias{nonest.basis}
\alias{nonest.basis.SemnovaEffect}
\alias{qr.SemnovaEffect}
\alias{model.matrix.SemnovaFormula}
\alias{terms.SemnovaFormula}
\alias{[.SemnovaFormula}
\title{SemnovaEffect class}
\usage{
\method{effSources}{SemnovaEffect}(mod)

\method{formula}{SemnovaEffect}(x, ...)

\method{family}{SemnovaEffect}(object, ...)

\method{model.frame}{SemnovaEffect}(object, ...)

model.offset.SemnovaEffect(object, ...)

\method{model.matrix}{SemnovaEffect}(object, ...)

\method{expand}{model.frame.SemnovaEffect}(model, extras, envir = environment(formula(model)), na.expand = FALSE)

\method{getCall}{SemnovaEffect}(x, ...)

vcov.SemnovaEffect(object, complete = TRUE, ...)

\method{coefficients}{SemnovaEffect}(object, ...)

\method{coef}{SemnovaEffect}(object, ...)

nonest.basis(x, ...)

\method{nonest.basis}{SemnovaEffect}(x, ...)

\method{qr}{SemnovaEffect}(x, ...)

\method{model.matrix}{SemnovaFormula}(object, ...)

\method{terms}{SemnovaFormula}(x, ...)

\method{`[`}{SemnovaFormula}(x, i)
}
\description{
Some description
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expose}}{Read only. Returns the private structure of the object. For debugging purposes only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SemnovaEffect-get_model_matrix}{\code{SemnovaEffect$get_model_matrix()}}
\item \href{#method-SemnovaEffect-new}{\code{SemnovaEffect$new()}}
\item \href{#method-SemnovaEffect-effect}{\code{SemnovaEffect$effect()}}
\item \href{#method-SemnovaEffect-effSources}{\code{SemnovaEffect$effSources()}}
\item \href{#method-SemnovaEffect-clone}{\code{SemnovaEffect$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SemnovaEffect-get_model_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-SemnovaEffect-get_model_matrix}{}}}
\subsection{Method \code{get_model_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SemnovaEffect$get_model_matrix(data = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SemnovaEffect-new"></a>}}
\if{latex}{\out{\hypertarget{method-SemnovaEffect-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SemnovaEffect$new(mod)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SemnovaEffect-effect"></a>}}
\if{latex}{\out{\hypertarget{method-SemnovaEffect-effect}{}}}
\subsection{Method \code{effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SemnovaEffect$effect(focal.predictors, interval = c("ci", "sd"), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SemnovaEffect-effSources"></a>}}
\if{latex}{\out{\hypertarget{method-SemnovaEffect-effSources}{}}}
\subsection{Method \code{effSources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SemnovaEffect$effSources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SemnovaEffect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SemnovaEffect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SemnovaEffect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
