% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{summary,lgc-method}
\alias{summary,lgc-method}
\title{Printing the summary for an LGC object.}
\usage{
\S4method{summary}{lgc}(object, ...)
}
\arguments{
\item{object}{\code{lgc} object. The object to get a summary about.}

\item{...}{Additional arguments. Currently none supported.}
}
\description{
Printing the summary for an LGC object.
}
