% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lgc.R, R/LgcBayes.R
\name{summary.Lgc}
\alias{summary.Lgc}
\title{Wrapper method for lgc$summary()}
\usage{
\method{summary}{Lgc}(object, ...)

\method{summary}{Lgc}(object, ...)
}
\arguments{
\item{object}{Object of class Lgc.}

\item{...}{Arguments passed down to lgc$summary().}
}
\description{
Wrapper method for lgc$summary()

Wrapper method for lgc$summary()
}
