% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{summary.Power}
\alias{summary.Power}
\title{Wrapper method for power$summary()}
\usage{
\method{summary}{Power}(x, ...)
}
\arguments{
\item{x}{Object of class Power.}

\item{...}{Arguments passed down to power$summary().}
}
\description{
Wrapper method for power$summary()
}
