% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Power}
\alias{Power}
\title{Power class}
\description{
Some description
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{expose}}{Read only. Returns the private structure of the object. For debugging purposes only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Power-new}{\code{Power$new()}}
\item \href{#method-Power-get_results}{\code{Power$get_results()}}
\item \href{#method-Power-power_script_lgc}{\code{Power$power_script_lgc()}}
\item \href{#method-Power-power_script_semnova}{\code{Power$power_script_semnova()}}
\item \href{#method-Power-power_analysis_lgc}{\code{Power$power_analysis_lgc()}}
\item \href{#method-Power-power_analysis_semnova}{\code{Power$power_analysis_semnova()}}
\item \href{#method-Power-print}{\code{Power$print()}}
\item \href{#method-Power-summary}{\code{Power$summary()}}
\item \href{#method-Power-plot}{\code{Power$plot()}}
\item \href{#method-Power-clone}{\code{Power$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-new"></a>}}
\if{latex}{\out{\hypertarget{method-Power-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-Power-get_results}{}}}
\subsection{Method \code{get_results()}}{
Returns a list of detailed results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-power_script_lgc"></a>}}
\if{latex}{\out{\hypertarget{method-Power-power_script_lgc}{}}}
\subsection{Method \code{power_script_lgc()}}{
Creates a power script for the Lgc class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$power_script_lgc(..., out_file = NULL, print = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-power_script_semnova"></a>}}
\if{latex}{\out{\hypertarget{method-Power-power_script_semnova}{}}}
\subsection{Method \code{power_script_semnova()}}{
Creates a power script for the Semnova class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$power_script_semnova(..., out_file = NULL, print = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-power_analysis_lgc"></a>}}
\if{latex}{\out{\hypertarget{method-Power-power_analysis_lgc}{}}}
\subsection{Method \code{power_analysis_lgc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$power_analysis_lgc(..., data_syntax, sample_size, replications)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-power_analysis_semnova"></a>}}
\if{latex}{\out{\hypertarget{method-Power-power_analysis_semnova}{}}}
\subsection{Method \code{power_analysis_semnova()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$power_analysis_semnova(..., data_syntax, sample_size, replications)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-print"></a>}}
\if{latex}{\out{\hypertarget{method-Power-print}{}}}
\subsection{Method \code{print()}}{
Prints the power for each of the model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Power-summary}{}}}
\subsection{Method \code{summary()}}{
Prints the power for the hypotheses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Power-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the Lgc dummy object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$plot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Power-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Power-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Power$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
