% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{print.Lgc}
\alias{print.Lgc}
\title{Wrapper method for lgc$print()}
\usage{
\method{print}{Lgc}(x, ...)
}
\arguments{
\item{x}{Object of class Lgc.}

\item{...}{Arguments passed down to lgc$print().}
}
\description{
Wrapper method for lgc$print()
}
