% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{reliabilities}
\alias{reliabilities}
\title{Prints the reliabilities of the indicators of latent variables.}
\usage{
reliabilities(lgc)
}
\arguments{
\item{lgc}{Object of class Lgc.}
}
\description{
Prints the reliabilities of the indicators of latent variables.
}
