% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Semnova.R
\name{Semnova}
\alias{Semnova}
\alias{semnova}
\alias{semnova_dummy}
\alias{Effect.Semnova}
\alias{plot.Semnova}
\alias{coefficients.Semnova}
\alias{coef.Semnova}
\title{Semnova class}
\usage{
semnova(...)

semnova_dummy(...)

\method{Effect}{Semnova}(focal.predictors, mod, ...)

\method{plot}{Semnova}(mod, focal.predictors, ...)

\method{coefficients}{Semnova}(object, ...)

coef.Semnova(object, ...)
}
\description{
Some description
}
\details{
\code{semnova()} is a wrapper for the R6 class constructor
\code{Semnova$new()}. See \code{Semnova$new()} below for a list of arguments.
}
\keyword{internal}
\section{Super class}{
\code{\link[semnova:Lgc]{semnova::Lgc}} -> \code{Semnova}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Semnova-get_model_matrix}{\code{Semnova$get_model_matrix()}}
\item \href{#method-Semnova-new}{\code{Semnova$new()}}
\item \href{#method-Semnova-specify}{\code{Semnova$specify()}}
\item \href{#method-Semnova-specify_dummy}{\code{Semnova$specify_dummy()}}
\item \href{#method-Semnova-clone}{\code{Semnova$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="anova"><a href='../../semnova/html/Lgc.html#method-Lgc-anova'><code>semnova::Lgc$anova()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="belongs_to"><a href='../../semnova/html/Lgc.html#method-Lgc-belongs_to'><code>semnova::Lgc$belongs_to()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="compare"><a href='../../semnova/html/Lgc.html#method-Lgc-compare'><code>semnova::Lgc$compare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="estimate"><a href='../../semnova/html/Lgc.html#method-Lgc-estimate'><code>semnova::Lgc$estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="fit_measures"><a href='../../semnova/html/Lgc.html#method-Lgc-fit_measures'><code>semnova::Lgc$fit_measures()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_covariates_mmodel"><a href='../../semnova/html/Lgc.html#method-Lgc-get_covariates_mmodel'><code>semnova::Lgc$get_covariates_mmodel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_data"><a href='../../semnova/html/Lgc.html#method-Lgc-get_data'><code>semnova::Lgc$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_estimates"><a href='../../semnova/html/Lgc.html#method-Lgc-get_estimates'><code>semnova::Lgc$get_estimates()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_etas_mmodel"><a href='../../semnova/html/Lgc.html#method-Lgc-get_etas_mmodel'><code>semnova::Lgc$get_etas_mmodel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_facet_grid"><a href='../../semnova/html/Lgc.html#method-Lgc-get_facet_grid'><code>semnova::Lgc$get_facet_grid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_labels"><a href='../../semnova/html/Lgc.html#method-Lgc-get_labels'><code>semnova::Lgc$get_labels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="get_lav_syntax"><a href='../../semnova/html/Lgc.html#method-Lgc-get_lav_syntax'><code>semnova::Lgc$get_lav_syntax()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="has_covariate"><a href='../../semnova/html/Lgc.html#method-Lgc-has_covariate'><code>semnova::Lgc$has_covariate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="has_latent_covariate"><a href='../../semnova/html/Lgc.html#method-Lgc-has_latent_covariate'><code>semnova::Lgc$has_latent_covariate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="has_latent_dv"><a href='../../semnova/html/Lgc.html#method-Lgc-has_latent_dv'><code>semnova::Lgc$has_latent_dv()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="has_resid_cov"><a href='../../semnova/html/Lgc.html#method-Lgc-has_resid_cov'><code>semnova::Lgc$has_resid_cov()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="has_within"><a href='../../semnova/html/Lgc.html#method-Lgc-has_within'><code>semnova::Lgc$has_within()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="is_indicator"><a href='../../semnova/html/Lgc.html#method-Lgc-is_indicator'><code>semnova::Lgc$is_indicator()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="is_latent"><a href='../../semnova/html/Lgc.html#method-Lgc-is_latent'><code>semnova::Lgc$is_latent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="power_analysis"><a href='../../semnova/html/Lgc.html#method-Lgc-power_analysis'><code>semnova::Lgc$power_analysis()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="print"><a href='../../semnova/html/Lgc.html#method-Lgc-print'><code>semnova::Lgc$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="reliabilities"><a href='../../semnova/html/Lgc.html#method-Lgc-reliabilities'><code>semnova::Lgc$reliabilities()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="set_data"><a href='../../semnova/html/Lgc.html#method-Lgc-set_data'><code>semnova::Lgc$set_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="summary"><a href='../../semnova/html/Lgc.html#method-Lgc-summary'><code>semnova::Lgc$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="test"><a href='../../semnova/html/Lgc.html#method-Lgc-test'><code>semnova::Lgc$test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="semnova" data-topic="Lgc" data-id="tests_table"><a href='../../semnova/html/Lgc.html#method-Lgc-tests_table'><code>semnova::Lgc$tests_table()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Semnova-get_model_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-Semnova-get_model_matrix}{}}}
\subsection{Method \code{get_model_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Semnova$get_model_matrix(dat = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Semnova-new"></a>}}
\if{latex}{\out{\hypertarget{method-Semnova-new}{}}}
\subsection{Method \code{new()}}{
Semnova class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Semnova$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Semnova-specify"></a>}}
\if{latex}{\out{\hypertarget{method-Semnova-specify}{}}}
\subsection{Method \code{specify()}}{
Specifies a Semnova object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Semnova$specify(
  data,
  id,
  dv,
  indicators = NULL,
  withins = NULL,
  betweens = NULL,
  contrasts_arg = list(),
  resid_cov = list(),
  equal_resid_cov = list(),
  sphericity = list(),
  covariates = NULL,
  normalize_contrasts = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{tibble. Will be converted to tibble if not already a
tibble.}

\item{\code{id}}{Character vector. Variables from the data set that
uniquely identify each case.}

\item{\code{dv}}{Character. Variable from the data set that contains the
dependent variable.}

\item{\code{indicators}}{Character. Variable from the data set that
contains indicators measuring the latent dependent eta variables.}

\item{\code{withins}}{Character vector. Variables from the data set that
contain within-subject factors.}

\item{\code{betweens}}{Character vector. Variables from the data set that
contain between-subject factors.}

\item{\code{contrasts_arg}}{Named list of characters. Each element specifies
the contrast coding scheme for the within or between subject factors.
Names of the list elements indicate the factor.}

\item{\code{resid_cov}}{List of character vectors. Each character vector
contains manifest variables from the data set among which residual
covariances should be implemented. Residual covariances can vary.}

\item{\code{equal_resid_cov}}{List of character vectors. Each character vector
contains manifest variables from the data set among which residual
covariances should be implemented. Residual covariances are
constrained to be equal.}

\item{\code{sphericity}}{List of character vectors or formula. If list: Each
element contains factor names for which sphericity should be imposed;
If formula: Left-hand side is either \code{TRUE} or \code{FALSE}. If
left-hand side is \code{TRUE}, sphericty is assumed except for effect
names occuring on the right-hand side. If \code{FALSE}, sphericty is
NOT assumed except for effect names occuring on the right-hand side.}

\item{\code{covariates}}{Named list of character vectors. Each list element
(character vector) represents an covairate variable. The elements
of the character vector are manifest variables (i.e., indicators)
from the data set that measure the corresponding covariate. The list
names correspond to the names of the covariates.
E.g.:
\code{
covariates = list(
    covariate1 = c("covariate11", "covariate12"),
    covariate2 = c("covariate21", "covariate22"),
    covariate3 = c("covariate31", "covariate32")
)}}

\item{\code{normalize_contrasts}}{Logical. Indicates whether contrasts
should be scaled to length equal to one.}

\item{\code{...}}{Parameters passed to the Lgc Class constructor.}

\item{\code{compound_symmetry}}{Logical. Indicates whether compound symmetry
should be imposed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Semnova-specify_dummy"></a>}}
\if{latex}{\out{\hypertarget{method-Semnova-specify_dummy}{}}}
\subsection{Method \code{specify_dummy()}}{
Specifies a Semnova object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Semnova$specify_dummy(
  n_within = 1,
  n_between = 1,
  n_indicator = 1,
  n_latent_covariate = NULL,
  n_manifest_covariate = NULL,
  contrasts_arg_within = NULL,
  contrasts_arg_between = NULL,
  contrasts_arg_covariate = NULL,
  sphericity = NULL,
  resid_cov = NULL,
  equal_resid_cov = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Semnova-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Semnova-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Semnova$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
