% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgc.R
\name{test_hypothesis}
\alias{test_hypothesis}
\title{Tests a hypothesis.}
\usage{
test_hypothesis(lgc, hypothesis)
}
\arguments{
\item{lgc}{Object of class Lgc.}

\item{hypothesis}{Object of class Hypothesis.}
}
\description{
Tests a hypothesis.
}
